+++ src/main/java/org/affirm/domain/Loan.java
@@ -4,20 +4,22 @@ import com.opencsv.bean.CsvBindByName;
 import lombok.Data;
 import lombok.ToString;
 
+import java.math.BigDecimal;
+
 @Data
 @ToString
 public class Loan {
     @CsvBindByName(column = "interest_rate")
-    private double interestRate;
+    private float interestRate;
 
     @CsvBindByName(column = "amount")
-    private Integer amount;
+    private BigDecimal amount;
 
     @CsvBindByName(column = "id")
     private int id;
 
     @CsvBindByName(column = "default_likelihood")
-    private double defaultLikelihood;
+    private float defaultLikelihood;
 
     @CsvBindByName(column = "state")
     private String state;
